requirements.txt
gradio
google-generativeai
nltk
python-dotenv